+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/chrome_custom_tabs/ChromeSafariBrowserManager.java
@@ -68,11 +68,15 @@ public class ChromeSafariBrowserManager implements MethodChannel.MethodCallHandl
     extras.putSerializable("options", options);
     extras.putSerializable("menuItemList", (Serializable) menuItemList);
 
-    Boolean isSingleInstance = (Boolean) options.get("singleInstance");
+    Boolean isSingleInstance = (Boolean) Util.getOrDefault(options, "isSingleInstance", false);
+    Boolean isTrustedWebActivity = (Boolean) Util.getOrDefault(options, "isTrustedWebActivity", false);
     if (CustomTabActivityHelper.isAvailable(activity)) {
-      intent = new Intent(activity, !isSingleInstance ? ChromeCustomTabsActivity.class : ChromeCustomTabsActivitySingleInstance.class);
+      intent = new Intent(activity, !isSingleInstance ? 
+              (!isTrustedWebActivity ? ChromeCustomTabsActivity.class : TrustedWebActivity.class) :
+              (!isTrustedWebActivity ? ChromeCustomTabsActivitySingleInstance.class : TrustedWebActivitySingleInstance.class));
       intent.putExtras(extras);
-      if ((Boolean) options.get("noHistory")) {
+      Boolean noHistory = (Boolean) Util.getOrDefault(options, "noHistory", false);
+      if (noHistory) {
         intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
       }
       activity.startActivity(intent);