+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/chrome_custom_tabs/TrustedWebActivitySingleInstance.java
@@ -1,0 +1,7 @@
+package com.pichillilorenzo.flutter_inappwebview.chrome_custom_tabs;
+
+public class TrustedWebActivitySingleInstance extends TrustedWebActivity {
+
+  protected static final String LOG_TAG = "TrustedWebActivitySingleInstance";
+
+}