+++ example/lib/chrome_safari_browser_example.screen.dart
@@ -67,7 +67,8 @@ class _ChromeSafariBrowserExampleScreenState
                     options: ChromeSafariBrowserClassOptions(
                         android: AndroidChromeCustomTabsOptions(
                             shareState: CustomTabsShareState.SHARE_STATE_OFF,
-                            singleInstance: false,
+                            isSingleInstance: false,
+                            isTrustedWebActivity: false,
                             keepAliveEnabled: true),
                         ios: IOSSafariOptions(
                             dismissButtonStyle: