+++ lib/src/chrome_safari_browser/android/chrome_custom_tabs_options.dart
@@ -11,23 +11,31 @@ import '../../in_app_webview/android/in_app_webview_options.dart';
 ///This class represents all the Android-only [ChromeSafariBrowser] options available.
 class AndroidChromeCustomTabsOptions
     implements ChromeSafariBrowserOptions, AndroidOptions {
-  ///Set to `false` if you don't want the default share item to the menu. The default value is `true`.
+  ///Use `shareState` instead.
   @Deprecated('Use `shareState` instead')
   bool? addDefaultShareMenuItem;
 
   ///The share state that should be applied to the custom tab. The default value is [CustomTabsShareState.SHARE_STATE_DEFAULT].
+  ///
+  ///**NOTE**: Not available in a Trusted Web Activity.
   CustomTabsShareState shareState;
 
   ///Set to `false` if the title shouldn't be shown in the custom tab. The default value is `true`.
+  ///
+  ///**NOTE**: Not available in a Trusted Web Activity.
   bool showTitle;
 
   ///Set the custom background color of the toolbar.
   Color? toolbarBackgroundColor;
 
   ///Set to `true` to enable the url bar to hide as the user scrolls down on the page. The default value is `false`.
+  ///
+  ///**NOTE**: Not available in a Trusted Web Activity.
   bool enableUrlBarHiding;
 
   ///Set to `true` to enable Instant Apps. The default value is `false`.
+  ///
+  ///**NOTE**: Not available in a Trusted Web Activity.
   bool instantAppsEnabled;
 
   ///Set an explicit application package name that limits
@@ -41,10 +49,28 @@ class AndroidChromeCustomTabsOptions
   bool keepAliveEnabled;
 
   ///Set to `true` to launch the Android activity in `singleInstance` mode. The default value is `false`.
-  bool singleInstance;
+  bool isSingleInstance;
 
   ///Set to `true` to launch the Android intent with the flag `FLAG_ACTIVITY_NO_HISTORY`. The default value is `false`.
   bool noHistory;
+
+  ///Set to `true` to launch the Custom Tab as a Trusted Web Activity. The default value is `false`.
+  bool isTrustedWebActivity;
+
+  ///Sets a list of additional trusted origins that the user may navigate or be redirected to from the starting uri.
+  ///
+  ///**NOTE**: Available only in a Trusted Web Activity.
+  List<String> additionalTrustedOrigins;
+
+  ///Sets a display mode of a Trusted Web Activity.
+  ///
+  ///**NOTE**: Available only in a Trusted Web Activity.
+  TrustedWebActivityDisplayMode? displayMode;
+
+  ///Sets a screen orientation. This can be used e.g. to enable the locking of an orientation lock type.
+  ///
+  ///**NOTE**: Available only in a Trusted Web Activity.
+  TrustedWebActivityScreenOrientation screenOrientation;
 
   AndroidChromeCustomTabsOptions(
       {@Deprecated('Use `shareState` instead') this.addDefaultShareMenuItem,
@@ -55,8 +81,12 @@ class AndroidChromeCustomTabsOptions
       this.instantAppsEnabled = false,
       this.packageName,
       this.keepAliveEnabled = false,
-      this.singleInstance = false,
-      this.noHistory = false});
+      this.isSingleInstance = false,
+      this.noHistory = false,
+      this.isTrustedWebActivity = false,
+      this.additionalTrustedOrigins = const [],
+      this.displayMode,
+      this.screenOrientation = TrustedWebActivityScreenOrientation.DEFAULT});
 
   @override
   Map<String, dynamic> toMap() {
@@ -70,8 +100,12 @@ class AndroidChromeCustomTabsOptions
       "instantAppsEnabled": instantAppsEnabled,
       "packageName": packageName,
       "keepAliveEnabled": keepAliveEnabled,
-      "singleInstance": singleInstance,
-      "noHistory": noHistory
+      "isSingleInstance": isSingleInstance,
+      "noHistory": noHistory,
+      "isTrustedWebActivity": isTrustedWebActivity,
+      "additionalTrustedOrigins": additionalTrustedOrigins,
+      "displayMode": displayMode?.toMap(),
+      "screenOrientation": screenOrientation.toValue()
     };
   }
 
@@ -88,8 +122,20 @@ class AndroidChromeCustomTabsOptions
     options.instantAppsEnabled = map["instantAppsEnabled"];
     options.packageName = map["packageName"];
     options.keepAliveEnabled = map["keepAliveEnabled"];
-    options.singleInstance = map["singleInstance"];
+    options.isSingleInstance = map["isSingleInstance"];
     options.noHistory = map["noHistory"];
+    options.isTrustedWebActivity = map["isTrustedWebActivity"];
+    options.additionalTrustedOrigins = map["additionalTrustedOrigins"];
+    switch(map["displayMode"]["type"]) {
+      case "IMMERSIVE_MODE":
+        options.displayMode = TrustedWebActivityImmersiveDisplayMode.fromMap(map["displayMode"]);
+        break;
+      case "DEFAULT_MODE":
+      default:
+        options.displayMode = TrustedWebActivityDefaultDisplayMode();
+        break;
+    }
+    options.screenOrientation = map["screenOrientation"];
     return options;
   }