+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `singleInstance` option for Android `ChromeSafariBrowser` implementation
 - Added `onDownloadStartRequest` event and deprecated old `onDownloadStart` event
 - Added `shareState` Android option for `ChromeSafariBrowser` class
+- Added support for Android TWA (Trusted Web Activity)
 - Fixed missing `onZoomScaleChanged` call for `InAppBrowser` class
 - Fixed `requestImageRef` method always `null` on iOS
 - Fixed "applicationNameForUserAgent is not work in ios" [#525](https://github.com/pichillilorenzo/flutter_inappwebview/issues/525)